+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Custom fields to Job Descriptions
 - Add custom data to job description search results
+- Add international_country_code to phone number details
 
 ## [5.6.0] - 2023-04-27
 ### Added